fix: eliminate Knex/raw SQL shim and isolate DB exceptions (#115)
- Delete AdminService.rawQuery() (dead code)
- Convert PortalService.loadTenantChain() from recursive CTE to iterative loop
- Convert portal.ts KB sources query to MikroORM QueryBuilder
- Convert retrieval.ts COUNT query to em.count()
- Wrap analytics raw SQL in AnalyticsRepository class
- Wrap retrieval.ts vector search in VectorSearchRepository class
- Delete src/db.ts after all consumers migrated
- Update test mocks (analytics, rag-retrieval, portal-kb-routes)

All 634 tests pass.

Closes #115
