+++ tests/analytics.test.ts
@@ -2,14 +2,21 @@
  * Analytics module unit tests
  *
  * Tests analytics query functions from src/analytics.ts.
- * Mocks the shared db.js query function — no real DB connections.
+ * Mocks the ORM module to provide a fake Knex instance (no real DB connections).
  */
 
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 
-// Must mock db before importing analytics (static import resolution)
-vi.mock('../src/db.js', () => ({
-  query: vi.fn(),
+// Build a mock Knex raw function we can control per-test
+const mockRaw = vi.fn();
+
+// Mock the ORM to provide a fake EntityManager with getKnex()
+vi.mock('../src/orm.js', () => ({
+  orm: {
+    em: {
+      getKnex: () => ({ raw: mockRaw }),
+    },
+  },
 }));
 
 import {
@@ -20,9 +27,6 @@ import {
   getAdminTimeseriesMetrics,
   getAdminModelBreakdown,
 } from '../src/analytics.js';
-import { query } from '../src/db.js';
-
-const mockQuery = query as ReturnType<typeof vi.fn>;
 
 /** A fully-populated analytics summary row as the DB would return. */
 const FULL_SUMMARY_ROW = {
@@ -51,10 +55,10 @@ const ZERO_SUMMARY_ROW = {
 };
 
 describe('getAnalyticsSummary', () => {
-  beforeEach(() => mockQuery.mockReset());
+  beforeEach(() => mockRaw.mockReset());
 
   it('returns summary object with correct shape and values', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [FULL_SUMMARY_ROW] });
+    mockRaw.mockResolvedValueOnce({ rows: [FULL_SUMMARY_ROW] });
 
     const result = await getAnalyticsSummary('tenant-1', 24);
 
@@ -70,7 +74,7 @@ describe('getAnalyticsSummary', () => {
   });
 
   it('returns zeros when no traffic (empty result set values all zero)', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [ZERO_SUMMARY_ROW] });
+    mockRaw.mockResolvedValueOnce({ rows: [ZERO_SUMMARY_ROW] });
 
     const result = await getAnalyticsSummary('tenant-1');
 
@@ -81,36 +85,37 @@ describe('getAnalyticsSummary', () => {
   });
 
   it('uses rollup CTE when rollup=true', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [ZERO_SUMMARY_ROW] });
+    mockRaw.mockResolvedValueOnce({ rows: [ZERO_SUMMARY_ROW] });
 
     await getAnalyticsSummary('tenant-1', 24, true);
 
-    const calledSql = mockQuery.mock.calls[0][0] as string;
+    const calledSql = mockRaw.mock.calls[0][0] as string;
     expect(calledSql).toContain('subtenant_tree');
   });
 
   it('uses direct tenant filter when rollup=false (default)', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [ZERO_SUMMARY_ROW] });
+    mockRaw.mockResolvedValueOnce({ rows: [ZERO_SUMMARY_ROW] });
 
     await getAnalyticsSummary('tenant-1', 24, false);
 
-    const calledSql = mockQuery.mock.calls[0][0] as string;
+    const calledSql = mockRaw.mock.calls[0][0] as string;
     expect(calledSql).not.toContain('subtenant_tree');
-    expect(calledSql).toContain('tenant_id = $1');
+    // After placeholder conversion $1 becomes ?, so check for tenant_id = ?
+    expect(calledSql).toContain('tenant_id = ?');
   });
 
   it('propagates query errors', async () => {
-    mockQuery.mockRejectedValueOnce(new Error('DB down'));
+    mockRaw.mockRejectedValueOnce(new Error('DB down'));
     await expect(getAnalyticsSummary('tenant-1')).rejects.toThrow('DB down');
   });
 });
 
 describe('getTimeseriesMetrics', () => {
-  beforeEach(() => mockQuery.mockReset());
+  beforeEach(() => mockRaw.mockReset());
 
   it('returns array of buckets with correct shape', async () => {
     const bucketDate = new Date('2024-01-01T00:00:00Z');
-    mockQuery.mockResolvedValueOnce({
+    mockRaw.mockResolvedValueOnce({
       rows: [{
         bucket: bucketDate,
         requests: 10,
@@ -137,22 +142,22 @@ describe('getTimeseriesMetrics', () => {
   });
 
   it('returns empty array when no traffic data', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [] });
+    mockRaw.mockResolvedValueOnce({ rows: [] });
     const result = await getTimeseriesMetrics('tenant-1');
     expect(result).toEqual([]);
   });
 
   it('uses rollup CTE when rollup=true', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [] });
+    mockRaw.mockResolvedValueOnce({ rows: [] });
     await getTimeseriesMetrics('tenant-1', 24, 60, true);
-    const calledSql = mockQuery.mock.calls[0][0] as string;
+    const calledSql = mockRaw.mock.calls[0][0] as string;
     expect(calledSql).toContain('subtenant_tree');
   });
 
   it('returns multiple buckets in order', async () => {
     const d1 = new Date('2024-01-01T00:00:00Z');
     const d2 = new Date('2024-01-01T01:00:00Z');
-    mockQuery.mockResolvedValueOnce({
+    mockRaw.mockResolvedValueOnce({
       rows: [
         { bucket: d1, requests: 5, tokens: 100, cost_usd: 0.001, avg_latency_ms: 150, error_rate: 0, avg_overhead_ms: 3, avg_ttfb_ms: 60 },
         { bucket: d2, requests: 8, tokens: 200, cost_usd: 0.002, avg_latency_ms: 180, error_rate: 0, avg_overhead_ms: 4, avg_ttfb_ms: 70 },
@@ -166,16 +171,16 @@ describe('getTimeseriesMetrics', () => {
   });
 
   it('propagates query errors', async () => {
-    mockQuery.mockRejectedValueOnce(new Error('timeseries query fail'));
+    mockRaw.mockRejectedValueOnce(new Error('timeseries query fail'));
     await expect(getTimeseriesMetrics('tenant-1')).rejects.toThrow('timeseries query fail');
   });
 });
 
 describe('getModelBreakdown', () => {
-  beforeEach(() => mockQuery.mockReset());
+  beforeEach(() => mockRaw.mockReset());
 
   it('returns array of model usage with correct shape', async () => {
-    mockQuery.mockResolvedValueOnce({
+    mockRaw.mockResolvedValueOnce({
       rows: [{
         model: 'gpt-4o',
         requests: 75,
@@ -198,20 +203,20 @@ describe('getModelBreakdown', () => {
   });
 
   it('returns empty array when no model data', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [] });
+    mockRaw.mockResolvedValueOnce({ rows: [] });
     const result = await getModelBreakdown('tenant-1');
     expect(result).toEqual([]);
   });
 
   it('uses rollup CTE when rollup=true', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [] });
+    mockRaw.mockResolvedValueOnce({ rows: [] });
     await getModelBreakdown('tenant-1', 24, 10, true);
-    const calledSql = mockQuery.mock.calls[0][0] as string;
+    const calledSql = mockRaw.mock.calls[0][0] as string;
     expect(calledSql).toContain('subtenant_tree');
   });
 
   it('returns multiple models sorted by request count', async () => {
-    mockQuery.mockResolvedValueOnce({
+    mockRaw.mockResolvedValueOnce({
       rows: [
         { model: 'gpt-4o', requests: 100, error_rate: 0, avg_latency_ms: 250, total_tokens: 5000, estimated_cost_usd: 0.05 },
         { model: 'gpt-3.5-turbo', requests: 40, error_rate: 0.05, avg_latency_ms: 100, total_tokens: 1000, estimated_cost_usd: 0.001 },
@@ -225,7 +230,7 @@ describe('getModelBreakdown', () => {
   });
 
   it('propagates query errors', async () => {
-    mockQuery.mockRejectedValueOnce(new Error('model query fail'));
+    mockRaw.mockRejectedValueOnce(new Error('model query fail'));
     await expect(getModelBreakdown('tenant-1')).rejects.toThrow('model query fail');
   });
 });
@@ -233,41 +238,41 @@ describe('getModelBreakdown', () => {
 // ── Admin variants ──────────────────────────────────────────────────────────
 
 describe('getAdminAnalyticsSummary', () => {
-  beforeEach(() => mockQuery.mockReset());
+  beforeEach(() => mockRaw.mockReset());
 
   it('returns summary across all tenants', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [FULL_SUMMARY_ROW] });
+    mockRaw.mockResolvedValueOnce({ rows: [FULL_SUMMARY_ROW] });
     const result = await getAdminAnalyticsSummary(undefined, 24);
     expect(result.totalRequests).toBe(100);
-    const calledSql = mockQuery.mock.calls[0][0] as string;
+    const calledSql = mockRaw.mock.calls[0][0] as string;
     expect(calledSql).not.toContain('tenant_id =');
   });
 
   it('filters to one tenant when tenantId provided', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [ZERO_SUMMARY_ROW] });
+    mockRaw.mockResolvedValueOnce({ rows: [ZERO_SUMMARY_ROW] });
     await getAdminAnalyticsSummary('tenant-abc', 24);
-    const calledSql = mockQuery.mock.calls[0][0] as string;
+    const calledSql = mockRaw.mock.calls[0][0] as string;
     expect(calledSql).toContain('tenant_id');
   });
 
   it('propagates errors', async () => {
-    mockQuery.mockRejectedValueOnce(new Error('admin summary fail'));
+    mockRaw.mockRejectedValueOnce(new Error('admin summary fail'));
     await expect(getAdminAnalyticsSummary()).rejects.toThrow('admin summary fail');
   });
 });
 
 describe('getAdminTimeseriesMetrics', () => {
-  beforeEach(() => mockQuery.mockReset());
+  beforeEach(() => mockRaw.mockReset());
 
   it('returns empty array when no data', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [] });
+    mockRaw.mockResolvedValueOnce({ rows: [] });
     const result = await getAdminTimeseriesMetrics(undefined, 24, 60);
     expect(result).toEqual([]);
   });
 
   it('returns buckets for all tenants', async () => {
     const d = new Date();
-    mockQuery.mockResolvedValueOnce({
+    mockRaw.mockResolvedValueOnce({
       rows: [{ bucket: d, requests: 20, tokens: 800, cost_usd: 0.01, avg_latency_ms: 200, error_rate: 0, avg_overhead_ms: 5, avg_ttfb_ms: 80 }],
     });
     const result = await getAdminTimeseriesMetrics();
@@ -276,16 +281,16 @@ describe('getAdminTimeseriesMetrics', () => {
   });
 
   it('propagates errors', async () => {
-    mockQuery.mockRejectedValueOnce(new Error('admin ts fail'));
+    mockRaw.mockRejectedValueOnce(new Error('admin ts fail'));
     await expect(getAdminTimeseriesMetrics()).rejects.toThrow('admin ts fail');
   });
 });
 
 describe('getAdminModelBreakdown', () => {
-  beforeEach(() => mockQuery.mockReset());
+  beforeEach(() => mockRaw.mockReset());
 
   it('returns model breakdown across all tenants', async () => {
-    mockQuery.mockResolvedValueOnce({
+    mockRaw.mockResolvedValueOnce({
       rows: [{ model: 'gpt-4o', requests: 200, error_rate: 0.01, avg_latency_ms: 280, total_tokens: 10000, estimated_cost_usd: 0.1 }],
     });
     const result = await getAdminModelBreakdown(undefined, 24, 10);
@@ -294,13 +299,13 @@ describe('getAdminModelBreakdown', () => {
   });
 
   it('returns empty array when no data', async () => {
-    mockQuery.mockResolvedValueOnce({ rows: [] });
+    mockRaw.mockResolvedValueOnce({ rows: [] });
     const result = await getAdminModelBreakdown();
     expect(result).toEqual([]);
   });
 
   it('propagates errors', async () => {
-    mockQuery.mockRejectedValueOnce(new Error('admin model fail'));
+    mockRaw.mockRejectedValueOnce(new Error('admin model fail'));
     await expect(getAdminModelBreakdown()).rejects.toThrow('admin model fail');
   });
 });
